+++ src/main/java/io/github/mmbishop/gwttest/model/Context.java
@@ -24,6 +24,7 @@ public class Context {
 
     public String testName;
     public TestPhase testPhase;
+    public Class<? extends Exception> expectedExceptionClass;
     public Exception thrownException;
 
 }